Dropdown in generic Shared FE components
