+++ frontend/src/modules/Shared/EnvironmentGroupSelect/index.js
@@ -1,0 +1,2 @@
+export * from './EnvironmentTeamDatasetsDropdown';
+export * from './EnvironmentTeamDropdown';
