+++ frontend/src/modules/Shared/index.js
@@ -1,3 +1,4 @@
 export * from './Comments';
+export * from './EnvironmentGroupSelect';
 export * from './KeyValueTags';
 export * from './Stack';
